Fix pipeline bugs and add aya results for scripts 03-04
- configs/models.yaml: fix aya model org (CohereForAI → CohereLabs)
- scripts/03: add HF login, residual-only extraction for speed
- scripts/04: fix principal angles to use Vt_top basis (not SVD on 4096x4096),
  add silhouette skip-existing, store subspace_bases for fast reuse
- src/activations/extract.py: pass HF token to NNsight LanguageModel
- src/probing/linear_probe.py: fix cv_accuracy (was training-set accuracy),
  add cv metrics to saved npz, add skip-existing + atomic write retry
- src/probing/effective_rank.py: cast float16 to float32 before SVD
- src/visualization/heatmaps.py: clip silhouette negatives to 0, add
  plot_principal_angles_by_tier (2x2 grid by tier)
- src/evaluation/{asr,coherence,backtranslate}.py: per-generation coherence,
  NLLB back-translation, vllm WildGuard scoring
- generate_figures.py: standalone ASR/coherence figure generator
- figures/: updated ASR heatmap, coherence heatmap, aya representation figs

+++ src/probing/effective_rank.py
@@ -47,8 +47,8 @@ def compute_rank_profile(
     Returns:
         Dict with keys: singular_values, effective_rank, energy_spectrum.
     """
-    # Center the activations
-    centered = activations - activations.mean(axis=0, keepdims=True)
+    # Center the activations; cast to float32 since linalg.svd doesn't support float16
+    centered = (activations - activations.mean(axis=0, keepdims=True)).astype(np.float32)
     U, S, Vt = np.linalg.svd(centered, full_matrices=False)
     rank = effective_rank_at_threshold(S, threshold)
     energy = np.cumsum(S ** 2) / (np.sum(S ** 2) + 1e-12)
